format: refactor  widget
